+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This is used for recording Invader's changes. This changelog is based on
 - invader-build: Fixed subpermutations not using the main permutation's gain
 - invader-build: Fixed wind period in fog tags not defaulting to 1-1
 - invader-build: Fixed light tags' falloff/cutoff angle not clamping correctly
+- invader-build: Fixed particle radius animation not defaulting correctly
 - invader-extract: Fixed minimum velocity and potential and not being multiplied
   by tick rate